more intelligent sendRemoteNote 'which' option
can be \all or an index number. I can also imagine this to be a specific remote name in the future (as order of remotes will be different every time). If which is not \all and also no number, the note will not be played remotely. This could be a setup for a routing matrix for incoming MIDI events for example